requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
cloudinary==1.31.0
dj-database-url==1.2.0
Django==4.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.14
